[FIX] doc: Truncate 680-line example data
The page shows the first 15 lines of a CSV in raw form as an example.
Next it shows the same data rendered as a table. However, it was
displaying the entire 680-line source data. `csv-table` doesn't have a
`:lines:` directive. I copied and truncated the original file for use
in the documentation.