[REV] base: revert 3e77072fbf3c386e40cb677e548c089173ad5244
Revert commit
"[FIX] base: clearer message in case of conflict for no_gap sequences"

This commit changed the error returned in case of SQL conflict for no_gap
sequences, wrongly disabling the automatic retry (service/model.py) in case
of such errors.

Also adds a comment in the dedicated test to ensure the behavior is explained
and the automatic retry isn't broken once again by the same kind of change.

closes odoo/odoo#59829

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>